# docs/swagger.yaml
definitions:
  ncobase_common_resp.Exception:
    properties:
      code:
        description: Business code
        type: integer
      data:
        description: Response data
      errors:
        description: Validation errors
      message:
        description: Message
        type: string
      status:
        description: HTTP status
        type: integer
    type: object
  ncobase_common_types.JSON:
    additionalProperties: {}
    type: object
...
  structs.ReadAsset:
    properties:
      bucket:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      endpoint:
        type: string
      extras:
        $ref: '#/definitions/ncobase_common_types.JSON'
      id:
        type: string
      name:
        type: string
      object_id:
        type: string
      path:
        type: string
      size:
        type: integer
      storage:
        type: string
      tenant_id:
        type: string
      type:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  structs.UpdateAssetBody:
    properties:
      bucket:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      endpoint:
        type: string
      extras:
        $ref: '#/definitions/ncobase_common_types.JSON'
      id:
        type: string
      name:
        type: string
      object_id:
        type: string
      path:
        type: string
      size:
        type: integer
      storage:
        type: string
      tenant_id:
        type: string
      type:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
info:
  contact: {}
  description: Next-Gen Modular Framework
  title: Ncobase
  version: 0.1.0
paths:
...
  /v1/assets:
    get:
      description: List assets based on specified parameters.
      parameters:
      - in: query
        name: cursor
        type: string
      - description: validate:"gte=1,lte=100"
        in: query
        name: limit
        type: integer
      - in: query
        name: object_id
        required: true
        type: string
      - in: query
        name: storage
        type: string
      - in: query
        name: tenant_id
        required: true
        type: string
      - in: query
        name: type
        type: string
      - in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            items:
              $ref: '#/definitions/structs.ReadAsset'
            type: array
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      summary: List assets
      tags:
      - assets
    post:
      consumes:
      - multipart/form-data
      description: Create one or multiple assets.
      parameters:
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      - description: Object ID associated with the asset
        in: formData
        name: object_id
        type: string
      - description: Tenant ID associated with the asset
        in: formData
        name: tenant_id
        type: string
      - description: Additional properties associated with the asset (JSON format)
        in: formData
        name: extras
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/structs.ReadAsset'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      security:
      - Bearer: []
      summary: Create assets
      tags:
      - assets
  /v1/assets/{slug}:
    delete:
      description: Delete a specific asset.
      parameters:
      - description: Slug of the asset to delete
        in: path
        name: slug
        required: true
        type: string
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/structs.ReadAsset'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      security:
      - Bearer: []
      summary: Delete asset
      tags:
      - assets
    get:
      description: Get details of a specific asset.
      parameters:
      - description: Slug of the asset to retrieve
        in: path
        name: slug
        required: true
        type: string
      - description: Type of retrieval ('download' or 'stream')
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/structs.ReadAsset'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      summary: Get asset
      tags:
      - assets
    put:
      consumes:
      - multipart/form-data
      description: Update an existing asset.
      parameters:
      - description: Slug of the asset to update
        in: path
        name: slug
        required: true
        type: string
      - description: Asset details
        in: body
        name: asset
        required: true
        schema:
          $ref: '#/definitions/structs.UpdateAssetBody'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/structs.ReadAsset'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      security:
      - Bearer: []
      summary: Update asset
      tags:
      - assets
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"